% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_formals.R
\name{extract_formals}
\alias{extract_formals}
\title{Extract the formals from a \code{roxy_block} object}
\usage{
extract_formals(roxy_block)
}
\arguments{
\item{roxy_block}{A \code{roxy_block} object}
}
\value{
A \code{data.frame} with a row for each formal, its value, and the
function it belongs to
}
\description{
Extract the formals from a \code{roxy_block} object
}
