% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_package_name.R
\name{parse_package_name}
\alias{parse_package_name}
\title{Parse a package name from a \code{.tar.gz} file path}
\usage{
parse_package_name(path)
}
\arguments{
\item{path}{The file path to a \code{.tar.gz} package source archive}
}
\value{
A list with the package's name and version string
}
\description{
Parse a package name from a \code{.tar.gz} file path
}
