% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxy_block.R
\name{parse_roxy_block}
\alias{parse_roxy_block}
\title{Parse a \code{roxy_block} object}
\usage{
parse_roxy_block(roxy_block)
}
\arguments{
\item{roxy_block}{A \code{roxy_block} object, from 
link[roxygen2]{parse_package}}
}
\value{
A list of parsed information from a roxy_block object such as its
title, description, and formals
}
\description{
Parse a \code{roxy_block} object
}
