% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_package.R
\name{process_package}
\alias{process_package}
\title{Process an individual package}
\usage{
process_package(path)
}
\arguments{
\item{path}{The path to the package's source}
}
\value{
A \code{data.frame} of the package's formals
}
\description{
Processes a package from its source extracted on disk.
}
