% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_packages.R
\name{process_packages}
\alias{process_packages}
\title{Process packages}
\usage{
process_packages(path, out_path = NULL)
}
\arguments{
\item{path}{The path to look for \code{.tar.gz} files.}

\item{out_path}{If set, write resulting \code{data.frame} out at \code{out_path} as a
\code{CSV}.}
}
\value{
A \code{data.frame} unless \code{out_path} is specified, in which case a \code{CSV}
file is written at \code{out_path} and \code{out_path} is returned invisibly.
}
\description{
Processes a set of \code{.tar.gz} archives at the specified \code{path}.
}
