% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_package.R
\name{process_package}
\alias{process_package}
\title{Process an individual package}
\usage{
process_package(targz_path)
}
\arguments{
\item{targz_path}{Path to a \code{.tar.gz}}
}
\value{
A \code{data.frame} of the package's formals
}
\description{
Processes a package as a `.tar.gz`` file and returns the extracted formal
arguments and their default values.
}
